simpleTextEditor now extends EditorBase
