+++ examples/index.html
@@ -18,7 +18,7 @@
 	<link rel="stylesheet" href="assets/css/bootstrap-theme.css" media="screen" ></link>
 	<link rel="stylesheet" href="assets/css/main.css"></link>
 	<link rel="stylesheet" href="build/react-data-grid.css"></link>
-	<script src="http://cdnjs.cloudflare.com/ajax/libs/react/0.13.1/react.js"></script>
+	<script src="http://cdnjs.cloudflare.com/ajax/libs/react/0.13.1/react-with-addons.js"></script>
 	<script src="https://code.jquery.com/jquery-1.11.2.min.js"></script>
 
 	<script src="build/libs.js"></script>
